edwards25519: remove Go 1.12 compatibility hack
