+++ fe_generic.go
@@ -4,6 +4,8 @@
 
 package edwards25519
 
+import "math/bits"
+
 func feMulGeneric(v, x, y *fieldElement) {
 	x0 := x.l0
 	x1 := x.l1
@@ -176,6 +178,16 @@ func feSquareGeneric(v, x *fieldElement) {
 
 	*v = fieldElement{r00, r10, r20, r30, r40}
 	v.carryPropagate()
+}
+
+// madd64 returns ol + oh * 2⁶⁴ = lo + hi * 2⁶⁴ + a * b. That is, it multiplies
+// a and b, and adds the result to the split uint128 [lo,hi].
+func madd64(lo, hi, a, b uint64) (ol uint64, oh uint64) {
+	oh, ol = bits.Mul64(a, b)
+	var c uint64
+	ol, c = bits.Add64(ol, lo, 0)
+	oh, _ = bits.Add64(oh, hi, c)
+	return
 }
 
 // carryPropagate brings the limbs below 52 bits by applying the reduction
